+++ src/main/java/cn/elytra/mod/nomi_horizons/NomiHorizons.java
@@ -25,6 +25,7 @@ public class NomiHorizons {
     public static final Logger LOG = LogManager.getLogger(Tags.MOD_NAME);
 
     public static final String MOD_ID = Tags.MOD_ID;
+    public static final String VERSION = Tags.VERSION;
 
     public static final NomiHorizonsAPI API = new NomiHorizonsAPI();