+++ src/main/java/cn/elytra/mod/nomi_horizons/NomiHorizonsMixinLoader.java
@@ -4,12 +4,15 @@ import com.google.common.collect.Lists;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonSyntaxException;
 import com.google.gson.annotations.SerializedName;
+import net.minecraftforge.fml.common.versioning.ComparableVersion;
 import zone.rong.mixinbooter.ILateMixinLoader;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.CopyOption;
 import java.nio.file.Files;
+import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.util.List;
 
@@ -24,6 +27,9 @@ public class NomiHorizonsMixinLoader implements ILateMixinLoader {
         @SerializedName("__comment__")
         public String comment = "Hi! This is Nomi Horizons mixin configuration, which controls which hard-injecting features you are interested in. Set to false to disable certain mixins.";
 
+        @SerializedName("__version__")
+        public String version;
+
         public boolean setCellCapacityTo144 = true;
         public boolean useNonPhantomFluidTankWidgets = true;
         public boolean useCleanroomPardon = true;
@@ -36,7 +42,7 @@ public class NomiHorizonsMixinLoader implements ILateMixinLoader {
         var path = new File("config/nomi_horizons.mixins.json").toPath();
 
         if(Files.notExists(path)) {
-            config = new Config();
+            config = newConfig();
             try {
                 Files.write(path, gson.toJson(config).getBytes(StandardCharsets.UTF_8), StandardOpenOption.CREATE);
             } catch(IOException e) {
@@ -45,14 +51,43 @@ public class NomiHorizonsMixinLoader implements ILateMixinLoader {
         } else {
             try {
                 config = gson.fromJson(new String(Files.readAllBytes(path), StandardCharsets.UTF_8), Config.class);
+
+                // check config version
+                var modVersion = new ComparableVersion(NomiHorizons.VERSION);
+                var configVersion = config.version != null ? new ComparableVersion(config.version) : null;
+                if(configVersion == null || modVersion.compareTo(configVersion) > 0) {
+                    NomiHorizons.LOG.warn("nomi_horizons.mixins.json is out-dated!");
+                    try {
+                        // update the config version
+                        config.version = NomiHorizons.VERSION;
+                        // write it to the file, and the default values are also included in.
+                        Files.write(path, gson.toJson(config).getBytes(StandardCharsets.UTF_8), StandardOpenOption.WRITE);
+                        NomiHorizons.LOG.info("Updated new configuration options in nomi_horizons.mixins.json, have a look at it!");
+                    } catch(IOException e) {
+                        NomiHorizons.LOG.error("Failed to update nomi_horizons.mixins.json", e);
+                    }
+                }
             } catch(IOException e) {
                 NomiHorizons.LOG.error("Failed to read nomi_horizons.mixins.json, using default settings", e);
-                config = new Config();
+                config = newConfig();
             } catch(JsonSyntaxException e) {
                 NomiHorizons.LOG.error("Failed to parse nomi_horizons.mixins.json, using default settings", e);
-                config = new Config();
+                config = newConfig();
+                try {
+                    Files.move(path, path.getParent().resolve("nomi_horizons.mixins.json.bak"), StandardCopyOption.REPLACE_EXISTING);
+                    Files.write(path, gson.toJson(config).getBytes(StandardCharsets.UTF_8), StandardOpenOption.CREATE);
+                    NomiHorizons.LOG.info("A brand new nomi_horizons.mixins.json is generated, have a look at it!");
+                } catch(IOException ex) {
+                    NomiHorizons.LOG.error("Failed to fix nomi_horizons.mixins.json, using default settings", ex);
+                }
             }
         }
+    }
+
+    private static Config newConfig() {
+        var config = new Config();
+        config.version = NomiHorizons.VERSION;
+        return config;
     }
 
     @Override